Move Follow to Contact namespace
